Decouple new project metrics
